Add: signup form and logic

+++ webapp/src/actions/authActions.js
@@ -17,13 +17,45 @@ const authFailure = errors => ({
   payload: { errors },
 });
 
+export const resetAuthFormErrors = () => ({
+  type: userTypes.RESET_AUTH_FORM_ERRORS,
+});
+
 export const login = ({ email = '', password = '' }) => dispatch => {
   dispatch(authStart());
 
-  // send user data
   const userData = { email, password };
+
+  // send user data
   axios
     .post(`${BASE_URL}/login`, userData)
+    .then(response => {
+      const { token } = response.data;
+      localStorage.setItem('token', `Bearer ${token}`);
+      dispatch(authSuccess(token));
+      history.push('/');
+    })
+    .catch(error => {
+      dispatch(authFailure(error.response.data.errors));
+    });
+};
+
+export const signup = formData => dispatch => {
+  dispatch(authStart());
+
+  // send user data
+  const userData = {
+    email: formData.email || '',
+    password: formData.password || '',
+    confirmPassword: formData.confirmPassword || '',
+    username: formData.username || '',
+    firstname: formData.firstname || '',
+    lastname: formData.lastname || '',
+  };
+
+  // send user data
+  axios
+    .post(`${BASE_URL}/signup`, userData)
     .then(response => {
       const { token } = response.data;
       localStorage.setItem('token', `Bearer ${token}`);
